fix: class servies

+++ src/main/java/KidAttend/demo/dto/response/classroom/ClassResponse.java
@@ -19,5 +19,5 @@ public class ClassResponse {
 
     private TeacherResponse teacher;
 
-    private Integer currentStudents;
+    private Long currentStudents;
 }
